templates/flask/base/requirements.txt
Flask
python-dotenv
gunicorn
uv